WIP Adds source package matching to Alpine
Grype matching against the source package information will result in fewer
False Negatives/more True Positives for Alpine images.  This patch moves
and modifies the mathod that provides source package matching from DPKG
to a common method and updates both the DPKG and Alpine matchers to use it.

This patch is WIP and I'm not sure if its the right direction or if it even
works.  The source matching method was ported from DPKG matcher to the APK
matcher and I found out through trying to write the unit test there was more
supporting code that needed to be added.

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>